requirements.txt
pandas
python-docx
customtkinter
pytest
openpyxl
# pip install -r requirements.txt